refactor: extract common utilities
